perf: completely eliminated floating-point operations from the hot loop (v5.0.75)

+++ include/shared/Detector.h
@@ -15,7 +15,7 @@ public:
     FovDetector();
     ~FovDetector();
 
-    float Scan(const RoiConfig& cfg);
+    int Scan(const RoiConfig& cfg);
 
 private:
     HDC m_hdcScreen;
